backend/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
filelock==3.18.0
Flask==2.3.3
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==21.2.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.13.1
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
Pillow==10.0.0
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
six==1.17.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.5.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
